add command: dice, gacha, luck, music, list, pic, speech, weather
