+++ src/bot/commands.ts
@@ -41,443 +41,6 @@ export async function commandSelector(message: Message) {
       await DotBotFunctions.Debug.debug(message, content);
       break;
     }
-    case 'speech': {
-      if (!isEnableFunction(functionNames.GPT_WITHOUT_KEY)) {
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription(`機能が有効化されていません。`);
-
-        message.reply({ content: `機能が有効化されてないよ！(GPT_WITHOUT_KEY)`, embeds: [send] });
-        return;
-      }
-      const chat = content.join(' ');
-      await DotBotFunctions.Chat.speech(message, chat);
-      break;
-    }
-    case 'no-system-gpt': {
-      if (!isEnableFunction(functionNames.GPT_WITHOUT_KEY)) {
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription(`機能が有効化されていません。`);
-
-        message.reply({ content: `機能が有効化されてないよ！(GPT_WITHOUT_KEY)`, embeds: [send] });
-        return;
-      }
-      const chat = content.join(' ');
-      await DotBotFunctions.Chat.talk(message, chat, LITELLM_MODEL.GPT_4O, LiteLLMMode.NOPROMPT);
-      break;
-    }
-    case 'memory': {
-      await DotBotFunctions.Chat.setMemory(message);
-      break;
-    }
-    case 'pic': {
-      // eslint-disable-next-line no-constant-condition
-      if (!isEnableFunction(functionNames.GPT)) {
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription(`機能が有効化されていません。`);
-
-        message.reply({ content: `機能が有効化されてないよ！(picture)`, embeds: [send] });
-        return;
-      }
-      const chat = content.join(' ');
-      await DotBotFunctions.Chat.generatePicture(message, chat);
-      break;
-    }
-    case 'tenki': {
-      if (!isEnableFunction(functionNames.FORECAST)) {
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription(`機能が有効化されていません。`);
-
-        message.reply({ content: `機能が有効化されてないよ！(FORECAST)`, embeds: [send] });
-        return;
-      }
-      await DotBotFunctions.Forecast.weather(message, content);
-      break;
-    }
-    case 'luck': {
-      await DotBotFunctions.Gacha.pickOmikuji(message, content);
-      break;
-    }
-    case 'gacha': {
-      await DotBotFunctions.Gacha.pickGacha(message, content);
-      break;
-    }
-    case 'g': {
-      await DotBotFunctions.Gacha.pickGacha(message, content);
-      break;
-    }
-    case 'gp': {
-      await DotBotFunctions.Gacha.showPercent(message);
-      break;
-    }
-    case 'gl': {
-      await DotBotFunctions.Gacha.pickGacha(message, ['limit']);
-      break;
-    }
-    case 'give': {
-      const uid = content[0];
-      const iid = Number(content[1]);
-
-      await DotBotFunctions.Gacha.givePresent(message, uid, iid);
-      break;
-    }
-    case 'celo': {
-      await DotBotFunctions.Dice.celo(message);
-      break;
-    }
-    case 'celovs': {
-      await DotBotFunctions.Dice.celovs(message);
-      break;
-    }
-    /**
-     * 音楽をキューに追加する.
-     */
-    case 'play':
-    case 'pl': {
-      if (!isEnableFunction(functionNames.YOUTUBE)) {
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription(`機能が有効化されていません。`);
-
-        message.reply({ content: `機能が有効化されてないよ！(YOUTUBE)`, embeds: [send] });
-        return;
-      }
-      try {
-        if (!content || content.length === 0) {
-          const send = new EmbedBuilder().setColor('#ff0000').setTitle(`エラー`).setDescription(`URLが不正`);
-
-          message.reply({ content: `YoutubeのURLを指定して～！`, embeds: [send] });
-          return;
-        }
-
-        const url = content[0];
-        const channel = message.member?.voice.channel;
-
-        if (!channel) {
-          const send = new EmbedBuilder()
-            .setColor('#ff0000')
-            .setTitle(`エラー`)
-            .setDescription(`userのボイスチャンネルが見つからなかった`);
-
-          message.reply({ content: `ボイスチャンネルに入ってから使って～！`, embeds: [send] });
-          return;
-        }
-
-        await DotBotFunctions.Music.add(channel, url, message.author.id);
-      } catch (e) {
-        const error = e as Error;
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription([error.name, error.message, error.stack].join('\n'));
-
-        message.reply({
-          content: `ありゃ、何かで落ちたみたい…？よかったらりむくんに送ってみてね！`,
-          embeds: [send],
-        });
-        return;
-      }
-      break;
-    }
-    case 'search':
-    case 'sc': {
-      if (!isEnableFunction(functionNames.YOUTUBE)) {
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription(`機能が有効化されていません。`);
-
-        message.reply({ content: `機能が有効化されてないよ！(YOUTUBE)`, embeds: [send] });
-        return;
-      }
-
-      try {
-        if (!content || content.length === 0) {
-          const send = new EmbedBuilder().setColor('#ff0000').setTitle(`エラー`).setDescription(`URLが不正`);
-
-          message.reply({ content: `検索する単語を指定してね！`, embeds: [send] });
-          return;
-        }
-
-        const words = content.join(' ');
-        const channel = message.member?.voice.channel;
-
-        if (!channel) {
-          const send = new EmbedBuilder()
-            .setColor('#ff0000')
-            .setTitle(`エラー`)
-            .setDescription(`userのボイスチャンネルが見つからなかった`);
-
-          message.reply({ content: `ボイスチャンネルに入ってから使って～！`, embeds: [send] });
-          return;
-        }
-
-        await DotBotFunctions.Music.search(channel, words);
-      } catch (e) {
-        const error = e as Error;
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription([error.name, error.message, error.stack].join('\n'));
-
-        message.reply({
-          content: `ありゃ、何かで落ちたみたい…？よかったらりむくんに送ってみてね！`,
-          embeds: [send],
-        });
-        return;
-      }
-      break;
-    }
-    case 'interrupt':
-    case 'pi': {
-      await interrupt(message, content);
-      break;
-    }
-    case 'stop':
-    case 'st': {
-      await stop(message);
-      break;
-    }
-    case 'rem':
-    case 'rm': {
-      await rem(message, content);
-      break;
-    }
-    case 'pause': {
-      const gid = message.guild?.id;
-      if (!gid) {
-        return;
-      }
-      await DotBotFunctions.Music.pause(message.channel as VoiceBasedChannel);
-      break;
-    }
-    case 'list': {
-      if (!content || content.length === 0) {
-        const playlists = await DotBotFunctions.Music.getPlaylist(message.author.id);
-
-        if (playlists.length === 0) {
-          const send = new EmbedBuilder()
-            .setColor('#ff0000')
-            .setTitle('エラー: ')
-            .setDescription('プレイリストが登録されていない');
-
-          message.reply({ content: `見つからなかった…もう一回登録してみて～！`, embeds: [send] });
-          return;
-        }
-
-        const description = playlists
-          .map(
-            (p) =>
-              `登録名: ${p.name}\n> プレイリスト名: ${p.title} | ループ: ${p.loop ? 'ON' : 'OFF'} | シャッフル: ${
-                p.shuffle ? 'ON' : 'OFF'
-              }`
-          )
-          .join('\n');
-
-        const send = new EmbedBuilder().setColor('#00ffff').setTitle('プレイリスト:').setDescription(description);
-
-        message.reply({ content: `プレイリストの一覧だよ！`, embeds: [send] });
-
-        return;
-      }
-
-      switch (content[0]) {
-        case 'rm':
-        case 'rem': {
-          try {
-            const name = content[1];
-            const deleted = await DotBotFunctions.Music.removePlaylist(message.author.id, name);
-
-            if (!deleted) {
-              const send = new EmbedBuilder()
-                .setColor('#ff0000')
-                .setTitle('エラー:')
-                .setDescription('削除するプレイリスト名を取得できなかった');
-
-              message.reply({
-                content: `削除できなかったみたい…もう一度名前を確認してみて！`,
-                embeds: [send],
-              });
-              return;
-            }
-
-            const send = new EmbedBuilder().setColor('#00ffff').setTitle('プレイリスト: ').setDescription('削除完了');
-
-            message.reply({ content: `削除できたよ～！`, embeds: [send] });
-            break;
-          } catch (e) {
-            const error = e as Error;
-            const send = new EmbedBuilder()
-              .setColor('#ff0000')
-              .setTitle(`エラー`)
-              .setDescription([error.name, error.message, error.stack].join('\n'));
-
-            message.reply({
-              content: `ありゃ、何かで落ちたみたい…？よかったら下のメッセージをりむくんに送ってみてね！`,
-              embeds: [send],
-            });
-            return;
-          }
-        }
-        case 'reg':
-        case 'add': {
-          try {
-            const name = content[1];
-            const url = content[2];
-
-            const repository = new PlaylistRepository();
-            const p = await repository.get(message.author.id, name);
-
-            if (p) {
-              const send = new EmbedBuilder()
-                .setColor('#ff0000')
-                .setTitle(`エラー`)
-                .setDescription(`プレイリストが存在している`);
-              message.reply({
-                content: `もうすでに同じ名前があるみたい……先に消すか別の名前にして！`,
-                embeds: [send],
-              });
-              return;
-            }
-
-            const playlistFlag = ytpl.validateID(url);
-            const movieFlag = ytdl.validateURL(url);
-
-            if (playlistFlag) {
-              const pid = await ytpl.getPlaylistID(url);
-              const playlist = await ytpl(pid);
-
-              await repository.save({
-                name: name,
-                user_id: message.author.id,
-                title: playlist.title,
-                url: playlist.url,
-              });
-
-              const send = new EmbedBuilder()
-                .setColor('#ff9900')
-                .setTitle(`登録`)
-                .setDescription(`登録名: ${name}\nプレイリスト名: ${playlist.title}\nURL: ${url}`);
-              message.reply({ content: `以下の内容で登録したよ～！`, embeds: [send] });
-
-              return;
-            }
-
-            if (movieFlag) {
-              const movie = await ytdl.getInfo(url);
-
-              await repository.save({
-                name: name,
-                user_id: message.author.id,
-                title: movie.videoDetails.title,
-                url: movie.videoDetails.video_url,
-              });
-
-              const send = new EmbedBuilder()
-                .setColor('#ff9900')
-                .setTitle(`登録`)
-                .setDescription(`登録名: ${name}\n動画名: ${movie.videoDetails.title}\nURL: ${url}`);
-              message.reply({ content: `以下の内容で登録したよ～！`, embeds: [send] });
-
-              return;
-            }
-
-            const send = new EmbedBuilder()
-              .setColor('#ff0000')
-              .setTitle(`エラー`)
-              .setDescription(`プレイリストor動画のURLではない`);
-            message.reply({ content: `プレイリストが非公開かも？`, embeds: [send] });
-
-            return false;
-          } catch (e) {
-            const error = e as Error;
-            const send = new EmbedBuilder()
-              .setColor('#ff0000')
-              .setTitle(`エラー`)
-              .setDescription([error.name, error.message, error.stack].join('\n'));
-
-            message.reply({
-              content: `ありゃ、何かで落ちたみたい…？よかったらりむくんに送ってみてね！`,
-              embeds: [send],
-            });
-            return;
-          }
-          break;
-        }
-        case 'loop':
-        case 'lp': {
-          const name = content[1];
-          const state = content[2].toUpperCase() === 'ON';
-
-          const repository = new PlaylistRepository();
-          const p = await repository.get(message.author.id, name);
-
-          if (!p) {
-            return;
-          }
-
-          await repository.save({ id: p.id, loop: Number(state) });
-          const send = new EmbedBuilder()
-            .setColor('#ff9900')
-            .setTitle(`更新`)
-            .setDescription(`更新名: ${name}\nループ: ${state ? 'ON' : 'OFF'}`);
-          message.reply({ content: `以下の内容で更新したよ～！`, embeds: [send] });
-          break;
-        }
-        case 'shuffle':
-        case 'sf': {
-          const name = content[1];
-          const state = content[2].toUpperCase() === 'ON';
-
-          const repository = new PlaylistRepository();
-          const p = await repository.get(message.author.id, name);
-
-          if (!p) {
-            return;
-          }
-
-          await repository.save({ id: p.id, shuffle: Number(state) });
-          const send = new EmbedBuilder()
-            .setColor('#ff9900')
-            .setTitle(`更新`)
-            .setDescription(`更新名: ${name}\nシャッフル: ${state ? 'ON' : 'OFF'}`);
-          message.reply({ content: `以下の内容で更新したよ～！`, embeds: [send] });
-          break;
-        }
-        default: {
-          break;
-        }
-      }
-      break;
-    }
-    case 'q': {
-      await queue(message);
-      break;
-    }
-    case 'silent':
-    case 'si': {
-      const channel = message.member?.voice.channel;
-      if (!channel) {
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription(`userのボイスチャンネルが見つからなかった`);
-
-        message.reply({ content: `ボイスチャンネルに入ってから使って～！`, embeds: [send] });
-        return;
-      }
-
-      await DotBotFunctions.Music.changeNotify(channel);
-      break;
-    }
     case 'mode': {
       const name = content[0];
       const channel = message.member?.voice.channel;
